Move error.hpp back
