+++ include/framework/everest.hpp
@@ -13,7 +13,7 @@
 #include <everest/exceptions.hpp>
 
 #include <utils/config.hpp>
-#include <utils/error/error.hpp>
+#include <utils/error.hpp>
 #include <utils/error/error_manager.hpp>
 #include <utils/mqtt_abstraction.hpp>
 #include <utils/types.hpp>
