+++ lib/error/error_type_map.cpp
@@ -3,7 +3,7 @@
 
 #include <utils/error/error_type_map.hpp>
 
-#include <utils/error/error.hpp>
+#include <utils/error.hpp>
 #include <utils/error/error_exceptions.hpp>
 #include <utils/yaml_loader.hpp>
